Reuse getEffectiveOtelStreamDataset
