+++ x-pack/platform/plugins/shared/fleet/server/services/agent_policies/package_policies_to_agent_inputs.ts
@@ -24,7 +24,6 @@ import type {
 import { DEFAULT_OUTPUT } from '../../constants';
 import { pkgToPkgKey } from '../epm/registry';
 import {
-  DATASET_VAR_NAME,
   DATA_STREAM_TYPE_VAR_NAME,
   FLEET_ENDPOINT_PACKAGE,
   GLOBAL_DATA_TAG_EXCLUDED_INPUTS,
@@ -39,6 +38,8 @@ import {
   packagePolicyInputAllowsUndefinedDataStreamType,
   getInputEffectiveName,
 } from '../../../common/services';
+
+import { getEffectiveOtelStreamDataset } from './get_effective_otel_stream_dataset';
 
 const isPolicyEnabled = (packagePolicy: PackagePolicy) => {
   return packagePolicy.enabled && packagePolicy.inputs && packagePolicy.inputs.length;
@@ -197,14 +198,11 @@ export const getFullInputStreams = (
               }
 
               if (input.type === OTEL_COLLECTOR_INPUT_TYPE) {
-                const datasetVar = stream.vars?.[DATASET_VAR_NAME]?.value;
                 // Replace policy output dataset verbatim (no .otel append); EPM templates use registry dataset + isOtelInputType separately.
-                if (datasetVar) {
-                  fullStream.data_stream = {
-                    ...fullStream.data_stream,
-                    dataset: datasetVar,
-                  };
-                }
+                fullStream.data_stream = {
+                  ...fullStream.data_stream,
+                  dataset: getEffectiveOtelStreamDataset(stream),
+                };
 
                 const useAPMVar = stream.vars?.[USE_APM_VAR_NAME]?.value;
                 if (useAPMVar !== undefined) {
